requirements.txt
# Required for YAML intent manifests
PyYAML>=6.0
